be/requirements.txt
eventlet==0.33.3
Flask==2.3.2
Flask-Cors==3.0.10
flask-serial==1.1.0
Flask-SocketIO==5.3.4
mysql-connector-python==8.0.33
python-socketio==5.8.0
